Split classes
Updated to two files. 1 for the agent class and 1 for function classes
